updated table header and added filtering column
